Quote $LDFLAGS and $CFLAGS paths on Windows
Currently, there is a bug on Windows regarding compiling Ashton with
rake. The problem is that if a full path to Ashton directory contains
spaces (for example, if Ashton's parent directory is "Documents
And Settings"), `rake compile` fails to compile anything.

Fix this by quotting flag paths that are appended to $LDFLAGS and
$CFLAGS variables.

Signed-off-by: Kyrylo Silin <kyrylosilin@gmail.com>